Issue: SCSAS-7253 libamxo build failed because when_true_status() macro is redefined.
From will.tallentire@consult.red :

I am building libamxo with the latest libamxc and there is a problem.

https://gitlab.com/prpl-foundation/components/ambiorix/libraries/libamxc/-/commit/4d6e1a64777ea59bc224a6105b2c8f3db02af24b
```
-#define when_true_status(x, l, c) if(x) { c; goto l; }
+#define when_true_status(x, l, c) if((x)) { c; goto l; }
```

This change causes a libamxo failure due to the following in libmxo/include_priv/amxo_parser_priv.h
```
/#include <amxc/amxc_macros.h>

/#define when_true_status(x, l, c) if(x) { c; goto l; }
```
This redefines the when_true_status macro, which is not required due to including from amxc_macros.h.

It used to work simply because the macro was identical so wasn't a redefinition.
Adding the extra brackets on if((x)) has made them different.

Can the unnecessary #define be removed from libmxo/include_priv/amxo_parser_priv.h?

GitOrigin-RevId: 4d7d139a702ab02b0fd11eecd5cf1f6f4d05aa66